Cleaner separation between validation and test.
This means that the validation and test mixins are now not based on the dataset, but rather on the intend:
    - Validation: Tests ran during training
    - Test: Tests ran separate from training

In order to achieve this, we removed the validation method on the engine, as it should now never be ran manually.
Instead, you can now specify which dataset to use for testing.
Technically, you can also alter which dataset you use for training/validation,
but as of now, this is not exposed to the engine, and requires manual tweaking of the ``engine.mixin_data``.

Finally, we also allow the end-users to specify the names of the ``Test_LoopMixin`` infer/post functions.
This allows to use the same mixin for testing and validation, and simply specify different function names if different functionality is required.